requirements.txt
coverage==6.5.0
coveralls==3.3.1
google-cloud-storage==3.5.0
jsonschema==4.23.0
numpy==1.26.0
pandas==2.2.3
pytest==7.2.2
pytest-cov==2.10.0
pyyaml==6.0.1
requests==2.32.4
requests-mock==1.8.0
retry==0.9.2
unidecode==1.4.0
